+++ waves/tests/test_latinhypercube.py
@@ -55,10 +55,17 @@ class TestLatinHypercube:
                 pass
 
     generate_input = {
-        "good schema":
+        "good schema 5x2":
             {'num_simulations': 5,
              'parameter_1': {'distribution': 'norm', 'loc': 50, 'scale': 1},
-             'parameter_2': {'distribution': 'norm', 'loc': -50, 'scale': 1}}
+             'parameter_2': {'distribution': 'norm', 'loc': -50, 'scale': 1}},
+        "good schema 2x1":
+            {'num_simulations': 2,
+             'parameter_1': {'distribution': 'norm', 'loc': 50, 'scale': 1}},
+        "good schema 1x2":
+            {'num_simulations': 1,
+             'parameter_1': {'distribution': 'norm', 'loc': 50, 'scale': 1},
+             'parameter_2': {'distribution': 'norm', 'loc': -50, 'scale': 1}},
     }
 
     @pytest.mark.unittest
@@ -69,14 +76,13 @@ class TestLatinHypercube:
         parameter_names = [key for key in parameter_schema.keys() if key != 'num_simulations']
         TestGenerate = LatinHypercube(parameter_schema)
         TestGenerate.generate()
-        values_array = TestGenerate.parameter_study.sel(parameter_data='values').to_array().values
-        quantiles_array = TestGenerate.parameter_study.sel(parameter_data='quantiles').to_array().values
+        values_array = TestGenerate.parameter_study['values'].values
+        quantiles_array = TestGenerate.parameter_study['quantiles'].values
         assert values_array.shape == (parameter_schema['num_simulations'], len(parameter_names))
-        assert numpy.all(value > 0 for value in values_array[:, 0])
-        assert numpy.all(value < 0 for value in values_array[:, 1])
         assert quantiles_array.shape == (parameter_schema['num_simulations'], len(parameter_names))
         # Verify that the parameter set name creation method was called
         assert TestGenerate.parameter_set_names == [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
         # Check that the parameter set names are correctly populated in the parameter study Xarray Dataset
-        parameter_set_names = list(TestGenerate.parameter_study.keys())
-        assert parameter_set_names == [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
+        expected_set_names = [f"parameter_set{num}" for num in range(parameter_schema['num_simulations'])]
+        parameter_set_names = list(TestGenerate.parameter_study['parameter_sets'])
+        assert numpy.all(parameter_set_names == expected_set_names)